rename RightSideBar to PinInformation and restructure subfolder
