+++ src/components/PinInformation/BoxInfo/BoxInfo.module.css
