+++ src/components/PinInformation/BoxList/BoxList.module.css
